Update all crate api paths used by the frontend to use the file safe crate name
This is currently foo~bar for a crate named foo/bar, but the ~ is
configurable. It was chosen because it can be inserted into a URL
without percent encoding, doesn't have another special meaning there,
and isn't a valid crate name so I can losslessly convert to and from
the actual crate name.